function score: fix cast in Gaussian decay function
Also fix the test
FunctionScoreTests#simpleWeightedFunctionsTestWithRandomWeightsAndRandomCombineMode
which sometimes failed due to rounding issues. Make sure
only floats are returned as scores to assure ratio of
expected and returned score is 1.0f.

+++ src/main/java/org/elasticsearch/index/query/functionscore/gauss/GaussDecayFunctionParser.java
@@ -40,7 +40,7 @@ public class GaussDecayFunctionParser extends DecayFunctionParser {
         public double evaluate(double value, double scale) {
             // note that we already computed scale^2 in processScale() so we do
             // not need to square it here.
-            return (float) Math.exp(0.5 * Math.pow(value, 2.0) / scale);
+            return Math.exp(0.5 * Math.pow(value, 2.0) / scale);
         }
 
         @Override
